# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Report System API",
        "description": "API documentation for the report system backend. Most school-scoped endpoints require an owner Sanctum token with the `permit:owner` ability.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/schools/{school}/marks": {
            "get": {
                "tags": [
                    "Marks"
                ],
                "summary": "List marks",
                "operationId": "c415d57a1d0e113721e283f0a16884cf",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "studentId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "schoolClassId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "termId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "subjectId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "academicYearId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marks retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Mark"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Marks"
                ],
                "summary": "Create a mark",
                "operationId": "cc7721019fd2f3c1851b31be5f02e61d",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "studentId",
                                    "subjectId",
                                    "schoolClassId",
                                    "studentClassEnrollmentId",
                                    "academicYearId",
                                    "termId",
                                    "classScore",
                                    "homeAssignmentScore",
                                    "projectScore",
                                    "classTestScore",
                                    "examScore"
                                ],
                                "properties": {
                                    "studentId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "subjectId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "schoolClassId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "studentClassEnrollmentId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "academicYearId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "termId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "classScore": {
                                        "type": "number",
                                        "example": 10,
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "homeAssignmentScore": {
                                        "type": "number",
                                        "example": 12,
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "projectScore": {
                                        "type": "number",
                                        "example": 14,
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "classTestScore": {
                                        "type": "number",
                                        "example": 13,
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "examScore": {
                                        "type": "number",
                                        "example": 70,
                                        "maximum": 100,
                                        "minimum": 0
                                    },
                                    "teacherId": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mark created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mark created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Mark"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/schools/{school}/marks/{mark}": {
            "put": {
                "tags": [
                    "Marks"
                ],
                "summary": "Update a mark",
                "operationId": "fcf6644d15f42130becd89b410726868",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "mark",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "classScore": {
                                        "type": "number",
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "homeAssignmentScore": {
                                        "type": "number",
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "projectScore": {
                                        "type": "number",
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "classTestScore": {
                                        "type": "number",
                                        "maximum": 15,
                                        "minimum": 0
                                    },
                                    "examScore": {
                                        "type": "number",
                                        "maximum": 100,
                                        "minimum": 0
                                    },
                                    "teacherId": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mark updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mark updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Mark"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Something went wrong."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Teacher": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "clientId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "schoolId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "staffNumber": {
                        "type": "string",
                        "example": "STF-1001"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0240000001"
                    }
                },
                "type": "object"
            },
            "Student": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "clientId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "schoolId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "schoolClassId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "firstName": {
                        "type": "string",
                        "example": "Akosua"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Mensah"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "admissionNumber": {
                        "type": "string",
                        "example": "ADM-1001"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date",
                        "example": "2012-04-15"
                    }
                },
                "type": "object"
            },
...
            "Subject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "schoolId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Mathematics"
                    }
                },
                "type": "object"
            },
...
            "Mark": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "studentId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "subjectId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "schoolClassId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "studentClassEnrollmentId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "academicYearId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "termId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "classScore": {
                        "type": "number",
                        "example": 10
                    },
                    "homeAssignmentScore": {
                        "type": "number",
                        "example": 12
                    },
                    "projectScore": {
                        "type": "number",
                        "example": 14
                    },
                    "classTestScore": {
                        "type": "number",
                        "example": 13
                    },
                    "continuousAssessmentScore": {
                        "type": "number"
                    },
                    "continuousAssessmentContribution": {
                        "type": "number"
                    },
                    "examScore": {
                        "type": "number",
                        "example": 70
                    },
                    "examContribution": {
                        "type": "number"
                    },
                    "totalScore": {
                        "type": "number"
                    },
                    "grade": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "gradeRemark": {
                        "type": "string",
                        "example": "Excellent",
                        "nullable": true
                    },
                    "teacherId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "student": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Student"
                            }
                        ],
                        "nullable": true
                    },
                    "subject": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Subject"
                            }
                        ],
                        "nullable": true
                    },
                    "schoolClass": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SchoolClass"
                            }
                        ],
                        "nullable": true
                    },
                    "academicYear": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AcademicYear"
                            }
                        ],
                        "nullable": true
                    },
                    "term": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Term"
                            }
                        ],
                        "nullable": true
                    },
                    "teacher": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Teacher"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token. Send as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}